requirements.txt
pdoc==12.*
flake8==6.0.0
pytest==7.1.*
openpyxl==3.1.2
matplotlib==3.8.3
seaborn==0.13.2
python-dotenv==1.0.1

streamlit==1.31.1

numpy
pillow==10.2.0

accelerate==0.27.2

torch==2.2.1
torchvision==0.17.1
transformers==4.38.2 
diffusers==0.26.3
invisible-watermark==0.2.0
tensorrt==8.6.1.post1